pyproject.toml
[project]
name = "electronics-distribution-network"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksim Lakovich",email = "maks.lakovich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django (>=5.2.2,<5.3)",
    "django-filter (>=25.2,<26.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^7.0.0"
mypy = "^1.18.2"

